requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2024.12.14
charset-normalizer==2.0.12
Django==3.2.25
djangorestframework==3.15.1
idna==3.10
importlib-resources==5.4.0
pytz==2024.2
requests==2.27.1
sqlparse==0.4.4
typing_extensions==4.1.1
urllib3==1.26.20
zipp==3.6.0
